refactor: scripts/env.sh

+++ scripts/env.sh
@@ -10,17 +10,20 @@ fi
 # INFORMATIONAL COMMANDS
 help() {
     cat <<EOF
-Usage: $SCRIPT <command>
- INFORMATIONAL COMMANDS
+Usage: $(basename "$0") <command>
+INFORMATIONAL COMMANDS
     help
         You are here.
- OPERATION COMMANDS
-    start <node_dir> <network_port>
-        Start the local node clusters.
-    stop
-        Stop the local node clusters.
-    clean
-        Clean the runtime files of the local node clusters.
+OPERATION COMMANDS
+    start <node_number>...
+        Start the local node clusters by node_number.
+        Start all the local node clusters if specified none.
+    stop <node_number>...
+        Stop the local node clusters by node_number.
+        Stop all the local node clusters if specified none.
+    clean <node_number>...
+        Clean the runtime files of the local node clusters by node_number.
+        Clean all the runtime files of the local node clusters if specified none.
 EOF
 
 }
@@ -28,12 +31,11 @@ EOF
 start() {
     export PATH=$WS/../bin:$PATH
 
-    local dir=$2
-    local port=$3
+    local dir="$WS/../config/$(ls "$WS/../config" | grep "$1"$)" || return
+
+    [ -f "$dir/pid_list" ] && echo "$(ls "$WS/../config" | grep "$1"$) is already running" && return
 
     cd "$dir"
-
-    pwd
 
     network run &
     pid_list[1]=$!
@@ -48,31 +50,122 @@ start() {
     crypto run &
     pid_list[6]=$!
 
-    cd -
+    echo -n "${pid_list[@]}" "" >> ./pid_list
 
-    echo -n "${pid_list[@]}" "" >> "$WS"/../pid_list
+    cd "$WS/.."
+
+    echo "$(basename "$dir") started"
+}
+
+startall() {
+    export PATH=$WS/../bin:$PATH
+
+    for node in $(ls "$WS/../config")
+    do
+        local dir="$WS/../config/$node"
+
+        [ -f "$dir/config.toml" ] || continue
+
+        [ -f "$dir/pid_list" ] && echo "$node is already running" && continue
+
+        cd "$dir"
+
+        network run &
+        pid_list[1]=$!
+        consensus run &
+        pid_list[2]=$!
+        executor run &
+        pid_list[3]=$!
+        storage run &
+        pid_list[4]=$!
+        controller run &
+        pid_list[5]=$!
+        crypto run &
+        pid_list[6]=$!
+
+        echo -n "${pid_list[@]}" "" >> ./pid_list
+
+        cd "$WS/.."
+
+        echo "$node started"
+        
+    done
 }
 
 stop() {
-    pid_list=$(cat "$WS"/../pid_list)
+    local dir="$WS/../config/$(ls "$WS/../config" | grep "$1"$)" || return
+    
+    [ -f "$dir/pid_list" ] || return
+
+    pid_list=$(cat "$dir/pid_list")
     for pid in ${pid_list[*]}; do
         kill "$pid"
     done
 
-    rm "$WS"/../pid_list
+    rm "$dir/pid_list"
+
+    echo "$(basename "$dir") stoped"
+}
+
+stopall() {
+    for node in $(ls "$WS/../config")
+    do
+        local dir="$WS/../config/$node"
+        [ -f "$dir/pid_list" ] || continue
+
+        pid_list=$(cat "$dir/pid_list")
+        for pid in ${pid_list[*]}; do
+            kill "$pid"
+        done
+
+        rm "$dir/pid_list"
+
+        echo "$node stoped"
+        
+    done
 }
 
 
 clean() {
-    if [ -f "$WS"/../pid_list ]; then
-        stop
+    local dir="$WS/../config/$(ls "$WS/../config" | grep "$1"$)" || return
+    
+    [ -f "$dir/config.toml" ] || return
+
+    if [ -f "$dir/pid_list" ]
+    then
+        pid_list=$(cat "$dir/pid_list")
+        for pid in ${pid_list[*]}; do
+            kill "$pid"
+        done
+        rm "$dir/pid_list"
     fi
 
-    local dir=$2
-
     rm -rf "$dir"/data "$dir"/logs "$dir"/chain_data "$dir"/raft-data-dir "$dir"/overlord_wal
+
+    echo "$(basename "$dir") cleaned"
 }
 
+cleanall() {
+    for node in $(ls "$WS/../config")
+    do
+        local dir="$WS/../config/$node"
+        [ -f "$dir/config.toml" ] || continue
+
+        if [ -f "$dir/pid_list" ]
+        then
+            pid_list=$(cat "$dir/pid_list")
+            for pid in ${pid_list[*]}; do
+                kill "$pid"
+            done
+            rm "$dir/pid_list"
+        fi
+
+        rm -rf "$dir"/data "$dir"/logs "$dir"/chain_data "$dir"/raft-data-dir "$dir"/overlord_wal
+
+        echo "$node cleaned"
+        
+    done
+}
 
 parse_command() {
     local command="$1"
@@ -83,16 +176,45 @@ parse_command() {
         ;;
 
     start)
-        start "$@"
+        if [ "$#" -gt 1 ]
+        then
+            for i in "$@"
+            do
+                expr $i + 1 >/dev/null 2>&1 || continue
+                start "$i"
+            done
+        else
+            startall
+        fi
         exit 0
         ;;
 
     stop)
-        stop
+        if [ "$#" -gt 1 ]
+        then
+            for i in "$@"
+            do
+                expr $i + 1 >/dev/null 2>&1 || continue
+                stop "$i"
+            done
+        else
+            stopall
+        fi
+        exit 0
         ;;
 
     clean)
-        clean "$@"
+        if [ "$#" -gt 1 ]
+        then
+            for i in "$@"
+            do
+                expr $i + 1 >/dev/null 2>&1 || continue
+                clean "$i"
+            done
+        else
+            cleanall
+        fi
+        exit 0
         ;;
 
     *)
